Fixed a small tags usage issue
