+++ lib/components/concert_tags_manager.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
+import '../APIfunctions/concertAPI.dart';
 
 class TagsUpdater {
 
   late List<String> tags;
+  late List<String> filteredTags;
   final changedNotifier = ValueNotifier<bool>(false);
 
   TagsUpdater() {
@@ -10,7 +12,20 @@ class TagsUpdater {
   }
 
   _init() {
-    tags = [];
+    tags = getTags();
+  }
+
+  void addFilteredTag(String entry) {
+    filteredTags.add(entry);
+  }
+
+  void removeFilteredTag(String entry) {
+    filteredTags.remove(entry);
+  }
+
+  void removeAllFilteredTags() {
+    filteredTags = [];
+    doUpdate();
   }
 
   void doUpdate() {
@@ -19,5 +34,14 @@ class TagsUpdater {
 
   void finUpdate() {
     changedNotifier.value = false;
+  }
+
+  List<String> getTags() {
+    List<String> toRet = [];
+    for(var entry in ConcertsAPI.getTags['tags']) {
+      toRet.add(entry);
+    }
+
+    return toRet;
   }
 }
